WIP refactor of LLM nonsense
